Adds more keys.
Adds more remote keys. support for different keys depends on the TV.
I also extended the interactive mode to include discrete inputs the
mappings are as follows.

    F1 - TV Source
    F2 - HDMI Source
    F3 - DVI Source
    F4 - DVR Source
    F5 - Digital TV Source
    F6 - Analog TV Source
    F7 - FM Radio Source
    F9 HDMI - 1 Source
    F10 HDMI - 2 Source
    F11 HDMI - 3 Source
    F12 HDMI - 4 Source
    CTRL+F1 - AV 1 Source
    CTRL+F2 - AV 2 Source
    CTRL+F3 - AV 3 Source
    CTRL+F5 - S Video 1 Source
    CTRL+F6 - S Video 2 Source
    CTRL+F7 - S Video 3 Source
    CTRL+F9 - Component 1 Source
    CTRL+F10 - Component 2 Source

I added the command line arg --key-help to the argument parser. I
supplied a default value to the positional key arguments. This is done
so that if --key-help is used and no positional arguments are passed
it will print off a full list of available keys. if keys are passed
it will print off a help for that key

basically the key help for a single key is to show the description of
that key. because there are keys that do not map out to what they
actually do. In the event someone forgets what a specific key command is
for.

Because I am running on Windows I added support for config file saving
and loading. as this threw an error when i ran the script.
I also did a little reorganization of the config file loading. this way
done when i was trying to track down the problem mentioned above. I
also added __future__.print_function for python 2 compatibility when
printing the help for the keys. The printing of the keys gets done before
a config file gets loaded. this is so that in the event the user has not
set up a config file yet the --key-help will still run.