Adds TV discovery
I added the ability for the library to discover any TV's using UPNP.
The UPNP detection sends out broadcast packets on all NIC's.
This process runs a different thread for each NIC so it will speed up
process. I have found that it can take up to 3 seconds for a TV to
reply to the SSDP packet.

I also added the user being able to select the TV they wish to use
if more then one is detected. the script will set up the config
automatically

the config values are

    host = found IP Address,
    id = PC MAC Address,
    name = 'samsungctl',
    description = PC Host Name,
    timeout = 0
    method = 'websocket' if year >= 2015 else 'legacy'
    port = 8001 if year >= 2015 else 55000

the automatic discovery process will take that 3 seconds, if the user
does not want to wait they will have to make a config file or pass the
settings from the command line.

I am able to pull all kinds of metrics from the UPNP SSDP request

    model
    panel type
    msg region
    size
    year
    max panel resolution
    series
    ip

I would say the serial number but i believe the serial number is
always the same no matter what tv is detected.