moves the setting of the auth event down one level
this is done so it gets set if there is an ms.channel.connect message
from the TV regardless if the token is present or not. as it turns out
the TV will not respond with a token if the one that was provided to it
when the websocket was created was correct.

Everything is now in place correctly for the websocket to function
properly with multiple TV's which is a good thing. It will dynamically
change tokens if the TV expires the one it has stored. this is all
seamless to the user.